Initial commit.
zeshy-specific syntax highlighting is enabled for all files with filetypes
".zy". Helpfully, such highlighting corrects numerous pressing issues in vim's
official zsh syntax highlighting -- particularly, the failure to match "(("-
and "))"-delimited test expressions.

Indeed, such failure causes use of the left bit shift operator "<<" in such
expressions to be catastrophically misinterpreted as a here-document prefix.
Naturally, vim then irreparably highlights the remainder of the current buffer
with the here-document string highlight group, effectively disabling *ALL*
syntax highlighting from then on. Yes, this is horrible. It's so horrible, in
fact, that we wrote this entire plugin merely to correct it.

That said, we expect to extend this plugin to support substantially improved
zeshy-specific syntax highlighting in the future. So, no significant harm done.